+++ docs/whats_new.rst
@@ -28,6 +28,7 @@ Enhancements
   regions (:pull:`361`, :issue:`346`).
 - The masking functions (e.g. :py:meth:`Regions.mask`) now warn if the `units` of the
   coordinates(``lat.attrs["units"]`` ) are given as "radians" (:issue:`279`).
+- Better error when passing a single region without wrapping it into a list or tuple (:issue:`372`).
 
 Deprecations
 ~~~~~~~~~~~~